requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
python-multipart==0.0.20
